% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/androidpublisher_objects.R
\name{SubscriptionPurchase}
\alias{SubscriptionPurchase}
\title{Google Play Developer API Objects 
Lets Android application developers access their Google Play accounts.}
\usage{
SubscriptionPurchase(autoRenewing = NULL, initiationTimestampMsec = NULL,
  validUntilTimestampMsec = NULL)
}
\arguments{
\item{autoRenewing}{Whether the subscription will automatically be renewed when it reaches its current expiry time}

\item{initiationTimestampMsec}{Time at which the subscription was granted, in milliseconds since the Epoch}

\item{validUntilTimestampMsec}{Time at which the subscription will expire, in milliseconds since the Epoch}
}
\value{
SubscriptionPurchase object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:26:19
filename: /Users/mark/dev/R/autoGoogleAPI/googleandroidpublisherv1.auto/R/androidpublisher_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
SubscriptionPurchase Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A SubscriptionPurchase resource indicates the status of a user's subscription purchase.
}
